rename helpers and add HelpersModule to AppModule
