+++ src/app/helpers/cookie-encryption/cookie-encryption.helper.ts
